Final version

+++ src/public/employee/delete.php
@@ -1,8 +1,10 @@
 <?php
+session_start();
 require_once __DIR__ . "/../../bootstrap/bootstrap.php";
 
 class EmployeeDeletePage extends CRUDPage
 {
+    use AdminAuthorization;
 
     protected function prepare(): void
     {
@@ -11,8 +13,17 @@ class EmployeeDeletePage extends CRUDPage
         $employeeId = filter_input(INPUT_POST, 'employeeId', FILTER_VALIDATE_INT);
         if (!$employeeId)
             throw new BadRequestException();
+        if ($employeeId === $_SESSION['id']) {
+            throw new ForbiddenException("You cannot delete yourself");
+        }
 
         //když poslal data
+        $keysStmt = PDOProvider::get()->prepare("SELECT r.room_id FROM room r JOIN `key` k ON r.room_id = k.room WHERE k.employee = :employeeId");
+        $keysStmt->execute(['employeeId' => $employeeId]);
+        if ($keysStmt->rowCount() !== 0) {
+            throw new ForbiddenException("Nemůžeš smazat zaměstnance, který má klíče");
+        }
+        $keys = $keysStmt->fetchAll(PDO::FETCH_UNIQUE);
         $success = Employee::deleteByID($employeeId);
 
         //přesměruj
@@ -23,7 +34,6 @@ class EmployeeDeletePage extends CRUDPage
     {
         return "";
     }
-
 }
 
 $page = new EmployeeDeletePage();
